pyproject.toml
[tool.poetry]
name = "tg-presentation-bot"
version = "0.1.0"
description = ""
authors = ["2rage <the2rage@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.6"
flask = "^3.0.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
